feat(config-flow): adopt PR #49 with B0/B1/B2 fixes (#49)
Lands the config-flow refactor + RepairsFlow contribution with the three
review blockers and one critical syntax bug resolved.

Brought in from PR #49:
- Extract HaghsDataUpdateCoordinator into custom_components/haghs/coordinator.py.
- Make CPU/RAM sensors optional in the config + options flow when PSI data is
  available; required otherwise. Schema deduplication via _schema_with_psi().
- New RepairsFlow (MissingFallbackSensorRepairFlow) so users can add fallback
  sensors via the Repairs UI if PSI becomes unavailable post-setup.
- async_setup_entry gates on PSI-or-fallbacks; raises ConfigEntryError +
  creates a fixable issue otherwise.
- async_migrate_entry to convert legacy text ignore-label values to label IDs.
- VersionInformation dataclass + IssueIds StrEnum in const.py.
- New fallback_missing strings + repair flow strings.

Fixes applied on top:
- B0: four except clauses in coordinator.py used Python 2 tuple syntax
  (`except A, B:`) which is a SyntaxError on Python 3.10+. Fixed all four to
  `except (A, B):` so the integration can actually load. Verified with
  ast.parse on every changed file.
- B1: strings.json used `invalid_config_entry` while runtime uses
  `IssueIds.FALLBACK_MISSING` ("fallback_missing"), and the fix-flow step id
  was `init` while repairs.py shows the form with step_id="confirm". Aligned
  strings.json with translations/en.json (which was already correct).
- B2: _migrate_ignore_label_value is now idempotent. An early-out via
  `label_registry.async_get_label(label_value)` skips entries whose ignore
  label is already a label ID. Without this guard, the next _CONFIG_VERSION
  bump would re-run the migration on already-migrated entries, miss the name
  lookup, and create a new label whose name equals the existing label's ID
  (silent data corruption).
- B3: extended v2.3_CHANGELOG.md with a "Config flow refactor + RepairsFlow"
  section documenting the new behaviour, the version bump (3.2), and the
  idempotency property of the migration.

Side-fix:
- tests/conftest.py: fixture had a bare `yield` with no teardown — replaced
  with `return` per PT022.

Bumps: VERSION=3, MINOR_VERSION=2.

Verified: ast.parse OK for all *.py, strings.json/en.json valid + matching
keys, ruff check clean. ruff format pass follows in a separate commit.
