+++ v2.3_CHANGELOG.md
@@ -12,6 +12,43 @@ Auto-detect under-voltage conditions on Raspberry Pi devices. When `binary_senso
 - `custom_components/haghs/__init__.py` — Added power supply check in `_async_calc_hardware()`, extended `_HardwareResult` with `p_power`, added recommendation in `_build_recommendations()`
 - `custom_components/haghs/const.py` — Added `REC_POWER_UNSTABLE` recommendation template
 
+### Config flow refactor + RepairsFlow for missing fallback sensors (#49)
+
+Builds on the `LabelSelector` change (#30) and reshapes setup so that CPU and
+RAM sensors are only required when Linux PSI data is unavailable. When PSI is
+available the user no longer has to pick percentage sensors at all; HAGHS uses
+the host's stall-time data directly.
+
+**What changed:**
+- `coordinator.py` — extracted `HaghsDataUpdateCoordinator` from `__init__.py`
+  into its own module. No scoring logic changes.
+- `config_flow.py` — `_schema_with_psi()` builds the schema with optional
+  CPU/RAM when PSI is available, required otherwise. Schema definitions
+  (`FALLBACK_SENSORS_SCHEMA`, `_BASE_SCHEMA`, `_EXTRA_OPTIONS_SCHEMA`)
+  deduplicated between setup and options flow. Switched to
+  `add_suggested_values_to_schema()` and `UnitOfTime.SECONDS`.
+- `repairs.py` (new) — `MissingFallbackSensorRepairFlow` lets users supply
+  CPU/RAM sensors through the Repairs UI if PSI becomes unavailable post-setup
+  (e.g. after a Docker host change), without uninstalling/reinstalling.
+- `__init__.py` — gates setup on PSI-or-fallbacks. If both are missing,
+  raises `ConfigEntryError(translation_key="fallback_missing")` and creates a
+  fixable repair issue. Adds `async_migrate_entry` to migrate legacy text
+  ignore-label values to label IDs.
+- `const.py` — added `VersionInformation` (frozen dataclass for ordered
+  version comparison) and `IssueIds` (StrEnum of issue identifiers).
+- `strings.json` / `translations/en.json` — added `fallback_missing` issue
+  and exception, plus the repair-flow `confirm` step.
+
+**Bumps:** `VERSION = 3`, `MINOR_VERSION = 2`.
+
+**Migration:** `_migrate_ignore_label_value` rewrites legacy text labels to
+label IDs and is idempotent — already-migrated entries are detected via
+`label_registry.async_get_label(label_value)` and skipped, so future
+`_CONFIG_VERSION` bumps cannot re-run the conversion and create duplicate
+labels.
+
+**Co-authored-by:** [@gustavakerstrom](https://github.com/gustavakerstrom)
+
 ## Bug Fixes
 
 ### Ignore Label: TextSelector replaced with LabelSelector (#30)
